Version 2.7.555 - Breaking change!
Warning:
========

Opening newer maps in old vym versions and saving them might cause data
loss!

In this version a new file format for notes and also headings is
introduced: Both are now in the XML file save as attributes of
<vymnote .../>  and <heading .../>

This makes encoding as CDATA obsolete and also allows properly
maintaining the history within the model of the map.

Old formats still can be read for backwards compatibility, but vym
versions < 2.5.518 will not correctly display maps.