bugfix: quell crash while checking a non-existent directory
Observed on Windows with MSVS 2019 an Qt 5.15.1:
When exporting to XML and the given target doesn't yet exist, the entry
list of it is empty. Then, accessing the first element of the empty list
lets the program crash.

Surprising that this seems to be not an issue for other systems.