Feature: Move a branch diagonally down (#19)
* Continued rewriting jobs

* Confluence agent improvements

* CA no longer uses external script to get page source and page details

* Creating new pages works in principle

* Create Confluence page: Uploading files

* Update of page works in principle

* Bugfix: exportLast for Confluence

* Getting user info works in principle

* Getting users from Confluence works in principle

* Renamed userdialog, introduce ConfluenceUser for results

* Cleanup, better use of attributes

* Bugfix: Exporting to Confluence natively

* Don't wait for keypress for expect_warnings_only

* Removed passing cdata to constructors in TreeModel

Only first column is used for headings in TreeEditor

Type in 2nd col was not used and there is no need to pass the first col
as parameter in constructors. THis is set later in setHeading

* Bugfix: exportLast of Markdown export

* Version 2.8.12

Replace macro Shift F10 with native code:

This allows to move a branch (e.g. a "timestamp" added to last log
entry) to become the sibling of its parent - a move diagonally down.