Setting Windows data-root directory
Don't add an additional data directory (usually "share") to the Windows
installation. Instead, put the demos, doc, exports, etc. folders
immediately in the directory where the executable lies.

The GNUInstallDirs CMake module honours the value of
CMAKE_INSTALL_DATAROOTDIR - if it was set in advance. That's why its
include needs to be after the set command.