Prepare saveState to use "complete" undo/redo commands
Selections are then part of the commands and don't need to be added with
hacks in undo() or redo(). First function to use this is

    VymModel::saveStateBranch()

which is called by the new scripting commands scroll() and unscroll().

General goal is to move all scripting commands refering to a branch from
VymModelWrapper to BranchWrapper. (Similar later for images, move these
commands to a yet to be created ImageWrapper)