Feature: link app icon as a mimetype icon for the hicolor default theme
On Linux/Unix systems cmake already installs the vym.png application
icon (what is referenced via the .desktop file) and a mime type definition
for `application/x-vym`. What is missing is the icon to use on `.vym`
files which are associated with this mime type. Instead of installing
the icon a second time, a relativ symlink is created referencing the app
icon.